# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * LLM provider configuration supplied by the user (BYOK)
 */
export interface LlmConfig {
  /** Provider identifier (e.g. openai, gemini, deepseek, openrouter, groq, mistral, together, minimax, custom) */
  provider: string;
  /** API key for the provider */
  apiKey: string;
  /** Model name to use */
  model: string;
  /** Optional custom base URL (required for provider=custom) */
  baseUrl?: string;
}

export interface FetchUrlBody {
  url: string;
}

export interface FetchUrlResponse {
  title: string;
  content: string;
  url: string;
}

export type LessonStatus = (typeof LessonStatus)[keyof typeof LessonStatus];

export const LessonStatus = {
  processing: "processing",
  ready: "ready",
  error: "error",
} as const;

export interface KeyConcept {
  term: string;
  definition: string;
}

export interface QuizQuestion {
  question: string;
  options: string[];
  correctIndex: number;
  explanation: string;
}

export interface Lesson {
  id: number;
  title: string;
  chapterText: string;
  summary: string;
  keyConcepts: KeyConcept[];
  quizQuestions: QuizQuestion[];
  status: LessonStatus;
  createdAt: string;
}

export interface CreateLessonBody {
  title: string;
  chapterText: string;
  llmConfig: LlmConfig;
}

export type ChatHistoryMessageRole =
  (typeof ChatHistoryMessageRole)[keyof typeof ChatHistoryMessageRole];

export const ChatHistoryMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatHistoryMessage {
  role: ChatHistoryMessageRole;
  content: string;
}

export interface ChatMessage {
  message: string;
  history: ChatHistoryMessage[];
  llmConfig: LlmConfig;
}

export interface ErrorResponse {
  error: string;
}