# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: lessons
    description: Lesson generation and management
  - name: content
    description: Content fetching utilities
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /fetch-url:
    post:
      operationId: fetchUrl
      tags: [content]
      summary: Fetch readable text content from a URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchUrlBody"
      responses:
        "200":
          description: Extracted text content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchUrlResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Could not extract content from URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List all lessons
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create a lesson from chapter text (triggers AI processing)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a lesson with full content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson with full content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete a lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lessons/{id}/chat:
    post:
      operationId: chatWithTutor
      tags: [lessons]
      summary: Send a message to the AI tutor for this lesson (SSE streaming)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessage"
      responses:
        "200":
          description: SSE stream of AI tutor response
          content:
            text/event-stream: {}
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lessons/{id}/chat-history:
    get:
      operationId: getLessonChatHistory
      tags: [lessons]
      summary: Get chat history for a lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatHistoryMessage"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    LlmConfig:
      type: object
      description: LLM provider configuration supplied by the user (BYOK)
      properties:
        provider:
          type: string
          description: Provider identifier (e.g. openai, gemini, deepseek, openrouter, groq, mistral, together, minimax, custom)
        apiKey:
          type: string
          description: API key for the provider
        model:
          type: string
          description: Model name to use
        baseUrl:
          type: string
          description: Optional custom base URL (required for provider=custom)
      required:
        - provider
        - apiKey
        - model
    FetchUrlBody:
      type: object
      properties:
        url:
          type: string
          format: uri
      required:
        - url
    FetchUrlResponse:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        url:
          type: string
      required:
        - title
        - content
        - url
    Lesson:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        chapterText:
          type: string
        summary:
          type: string
        keyConcepts:
          type: array
          items:
            $ref: "#/components/schemas/KeyConcept"
        quizQuestions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        status:
          type: string
          enum: [processing, ready, error]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - chapterText
        - summary
        - keyConcepts
        - quizQuestions
        - status
        - createdAt
    KeyConcept:
      type: object
      properties:
        term:
          type: string
        definition:
          type: string
      required:
        - term
        - definition
    QuizQuestion:
      type: object
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        explanation:
          type: string
      required:
        - question
        - options
        - correctIndex
        - explanation
    CreateLessonBody:
      type: object
      properties:
        title:
          type: string
        chapterText:
          type: string
        llmConfig:
          $ref: "#/components/schemas/LlmConfig"
      required:
        - title
        - chapterText
        - llmConfig
    ChatMessage:
      type: object
      properties:
        message:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatHistoryMessage"
        llmConfig:
          $ref: "#/components/schemas/LlmConfig"
      required:
        - message
        - history
        - llmConfig
    ChatHistoryMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error