# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Fetch readable text content from a URL
 */
export const FetchUrlBody = zod.object({
  url: zod.string().url(),
});

export const FetchUrlResponse = zod.object({
  title: zod.string(),
  content: zod.string(),
  url: zod.string(),
});

/**
 * @summary List all lessons
 */
export const ListLessonsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  chapterText: zod.string(),
  summary: zod.string(),
  keyConcepts: zod.array(
    zod.object({
      term: zod.string(),
      definition: zod.string(),
    }),
  ),
  quizQuestions: zod.array(
    zod.object({
      question: zod.string(),
      options: zod.array(zod.string()),
      correctIndex: zod.number(),
      explanation: zod.string(),
    }),
  ),
  status: zod.enum(["processing", "ready", "error"]),
  createdAt: zod.date(),
});
export const ListLessonsResponse = zod.array(ListLessonsResponseItem);

/**
 * @summary Create a lesson from chapter text (triggers AI processing)
 */
export const CreateLessonBody = zod.object({
  title: zod.string(),
  chapterText: zod.string(),
  llmConfig: zod
    .object({
      provider: zod
        .string()
        .describe(
          "Provider identifier (e.g. openai, gemini, deepseek, openrouter, groq, mistral, together, minimax, custom)",
        ),
      apiKey: zod.string().describe("API key for the provider"),
      model: zod.string().describe("Model name to use"),
      baseUrl: zod
        .string()
        .optional()
        .describe("Optional custom base URL (required for provider=custom)"),
    })
    .describe("LLM provider configuration supplied by the user (BYOK)"),
});

/**
 * @summary Get a lesson with full content
 */
export const GetLessonParams = zod.object({
  id: zod.coerce.number(),
});

export const GetLessonResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  chapterText: zod.string(),
  summary: zod.string(),
  keyConcepts: zod.array(
    zod.object({
      term: zod.string(),
      definition: zod.string(),
    }),
  ),
  quizQuestions: zod.array(
    zod.object({
      question: zod.string(),
      options: zod.array(zod.string()),
      correctIndex: zod.number(),
      explanation: zod.string(),
    }),
  ),
  status: zod.enum(["processing", "ready", "error"]),
  createdAt: zod.date(),
});

/**
 * @summary Delete a lesson
 */
export const DeleteLessonParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Send a message to the AI tutor for this lesson (SSE streaming)
 */
export const ChatWithTutorParams = zod.object({
  id: zod.coerce.number(),
});

export const ChatWithTutorBody = zod.object({
  message: zod.string(),
  history: zod.array(
    zod.object({
      role: zod.enum(["user", "assistant"]),
      content: zod.string(),
    }),
  ),
  llmConfig: zod
    .object({
      provider: zod
        .string()
        .describe(
          "Provider identifier (e.g. openai, gemini, deepseek, openrouter, groq, mistral, together, minimax, custom)",
        ),
      apiKey: zod.string().describe("API key for the provider"),
      model: zod.string().describe("Model name to use"),
      baseUrl: zod
        .string()
        .optional()
        .describe("Optional custom base URL (required for provider=custom)"),
    })
    .describe("LLM provider configuration supplied by the user (BYOK)"),
});

/**
 * @summary Get chat history for a lesson
 */
export const GetLessonChatHistoryParams = zod.object({
  id: zod.coerce.number(),
});

export const GetLessonChatHistoryResponseItem = zod.object({
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
});
export const GetLessonChatHistoryResponse = zod.array(
  GetLessonChatHistoryResponseItem,
);