# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatHistoryMessage } from "./chatHistoryMessage";
import type { LlmConfig } from "./llmConfig";

export interface ChatMessage {
  message: string;
  history: ChatHistoryMessage[];
  llmConfig: LlmConfig;
}