# lib/api-zod/src/generated/types/llmConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * LLM provider configuration supplied by the user (BYOK)
 */
export interface LlmConfig {
  /** Provider identifier (e.g. openai, gemini, deepseek, openrouter, groq, mistral, together, minimax, custom) */
  provider: string;
  /** API key for the provider */
  apiKey: string;
  /** Model name to use */
  model: string;
  /** Optional custom base URL (required for provider=custom) */
  baseUrl?: string;
}